pin and gpio fixes for WLAN_PD# --> WLAN_WAKE conversion
on CL4, GPIO_35 was initially hooked up to WLAN_PD# on the wireless
module.  we've never used that signal, and cforth arranged for it
to be high when running and asleep.

on C2 and later boards the pin was repurposed as WLAN_WAKE, and
hooked to an output of the wireless module.  in this mode, the
pin should be an input.

since the WLAN_PD# pin is pulled high on the wireless modules, it's
safe to configure GPIO_35 as an input all the time on all boards.

hopefully this completes this cycle of churn (c17ebbe, 207045b,
and 0ad3457).