CForth - relative paths are now based on the input file not pwd.
We now keep a stack of the absolute paths to the currently
active nested set of input files.  When opening a file with
a relative pathname, it is interpreted relative to the top
of that stack, i.e. the enclosing input file.  This is nearly
always the way you want it to work, so that "load files" can
name subordinate files relative to themselves, and you can
load such a file from "far away".

When the input is interactive or coming from a command line
string, the base for relative paths is still the current
working directory, so interactive use does the expected thing
of opening local files.

Absolute pathnames remain absolute, unchanged by this patch.

Created files are still relative to the current directory.

This change greatly simplies makefile rules that create .dic
files; it is no longer necessary to cd to the directory with
the source files, nor to copy the generated .dic file back
to the build location.  You can instead just load the top-level
load file in the far directory and everything it references
is automatically relative to it.

Change-Id: I510addbc1b7bd7e23c21f66c440c279afd3e9b17