No need to save/restore the floating stack pointer
The XFP user variable is the frame pointer for locals, not
the floating point stack pointer.  When entering/exiting
inner_interpreter(), there is no need to put the floating
point stack pointer in an externally-visible variable,
because we currently do not expose the floating stack to
things like execute_word().  The attempts to save/restore
the floating stack pointer in inner_interpreter() were
misguided, as they put it in the frame pointer variable.
One bad effect was crashes in kernel.dic if FLOATING was
enabled, because the XFP variable starts out uninitialized.