Embed target now allows system memory allocator
embed/targets.mk used to assume the use of mallocembed.o,
the rudimentary allocator that takes memory from the top
of the dictionary area.  That is appropriate for cases where
there is no external OS, but if there is an OS, it is better
to use its memory allocator.  The choice of allocator is now
left to the platform via src/app/*/targets.mk, which can use
mallocembed.o as before or can define getmem(), memfree(),
and memresize() some other way.

This change was triggered by the desire to use the system
allocator in the esp8266 version.