ESP8266 TCP - add dns-gethostbyname
For queries by name, returns `ERR_INPROGRESS`, and the callback is given address of IP address, or 0 if name was not found.

For queries where the name is a numeric IP address, returns `ERR_OK`, with IP filled in, and no callback.

An extern was used because the build was unhappy with `dns.h` and didn't have the patience to resolve it.

Test case, in station mode associated with an access point:

```
: ?.ipaddr  ( a.ip -- )  ?dup if  .ipaddr  else  ." NULL"  then  ;

: dns-cb  ( a.arg a.ip a.name -- )
   rot
   ." dns-cb arg " . cr
   swap
   ." dns-cb ip " ?.ipaddr cr
   ." dns-cb name " cscount type cr
;

create ip  4 allot  \ not used unless query is numeric

42 ' dns-cb ip c" google.com"            dns-gethostbyname . cr d# 1000 ms
43 ' dns-cb ip c" tooraweenah.com"       dns-gethostbyname . cr d# 1000 ms
44 ' dns-cb ip c" bogus.tooraweenah.com" dns-gethostbyname . cr d# 1000 ms
45 ' dns-cb ip c" 127.0.0.1"             dns-gethostbyname . cr ip ?.ipaddr cr
```