Fixed heap corruption with long [ifdef] lines
The problem was in ip_canonical(), which double-incremented
a pointer as it stepped through the word.  [else] calls
$canonical after parsing each word of the input to be
discarded.  If a single word exceeded half the size of
the input buffer, the double-increment would cause a write
past the end of the buffer, thus possibly smashing the heap.
This only showed up when compiling md5.fth from the OFW source.
That file contains an [ifdef] notdef block with an embedded
C comment block with a line of 72 asterisks.