Fix T16 compilation
A recent change broke the compilation of versions that use
T16 token encoding to save space.  That change increased
the default dictionary size to a larger value than T16
encoding can handle.  This fix ensures sets the default
dictionary size, when using T16, to the max value possible
for T16.