arm-xo-1.75: check that we don't use Thumb instructions
However, some distributions, such as Debian/armhf, ship with libgcc that uses
Thumb-2 instructions in their toolchains. The PJ1/Mohawk "security processor"
core has no support for that.

Let's add a sanity check, because CForth not working is a big deal (the
machine can't be unbricked without disassembling it).

Sadly, there doesn't seem to be a way to tell the linker to avoid
thumb-compiled binaries. Grepping for 16-bit instrucitons in our
disasembly does the trick for now.