arm-xo-1.75: default to using the cross-compiler
The native toolchain of Debian/armhf, and perhaps some other operating
systems, ships with libgcc that's no good for running on the "security
processor" PJ1/Mohawk core, because it uses Thumb-2 instructions.

Let's switch to the arm-none-eabi toolchain. It is available on both
Debian and Fedora regardless of what the native architecture is, so
it might be a pretty good default.