arm-ariel: add support for Dell Wyse 3020 machine
This adds support for using CForth on the MMP3-based Ariel board, as
used in Dell Wyse 3020. It starts up on the small core and loads OFW
in a manner equivalent to what is done on OLPC CL4.

Among the differencies from CL4 are: diffrent DRAM, generally less
hardware hooked to GPIOs (no special keys, etc.), different EC,
and different display controller. The only UART that's routed to actual
connector on board is UART3 on pins 51 and 52.

The machine originally comes with proprietary "WLoader" firmware.
Running Open Firmware on it would make it suck considerably less.