cforth: fix build with GCC 10
GCC 10 decided to turn on -fno-common by default [1] which fails the
linking in case the same symbol is defined by multiple object files.
We do trigger one such case, by accident:

  /usr/bin/ld: dictfile.o:src/cforth/dictfile.c:7:
      multiple definition of `variables';
      meta.o:src/cforth/meta.c:14: first defined here
  collect2: error: ld returned 1 exit status

Let's take this opportinity to make symbols that are of not of interest
to other object files static, at least for the two files that collided
here.

[1] https://gcc.gnu.org/gcc-10/porting_to.html