riscv-gd32vf103: add GD32VF103 target
This adds a very basic support for a MCU with STM32-compatible peripherals
and a 32-bit RISC-V core. Loosely based on based on arm-stm32f2013. The
peripheral driver library has a somewhat different interface though.

Tested by blinking a led on a Polos Alef R1 board, though my friends and
family didn't appreciate it nearly as much as I did.