Fixed GPIO problems
The SDK functions that I was using to setup the GPIOs were
incomplete, since they do not configure the pinmux.  The fix
was to switch to gpio_config(), which does all the right things.

One symptom was that, if you say "gpio13 gpio-is-output",
subsequent uses of "n gpio13 gpio-pin!" would not necessarily
do anything to the pin, because the pin would still be attached
to a different functional block.